+++ frameworks/DebugLocalizationPackage/Sources/DebugLocalizationCore/DebugLocalizationVersion.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public enum DebugLocalizationVersion {
-    public static let current = "0.2.0"
-}